Instructions in Main Menu and Shield in Level 1
In the main menu I added the instructions for the user to know how to
play the game, while I also added a shield in level 1 as a power up.
When the player collects 5 fires, the player will have a fire shield.
Upon hitting the enemy the user will loose that shield. The shield can
only be used once in that level